requirements.txt
python-telegram-bot==20.4
yt-dlp
aiohttp
ffmpeg-python
